Remove "Ultra Tech" template from meta-traits folder, split into 2 actual libraries
The confusingly named file is just sitting in a folder, and is a template file rather than a library file. It is the only thing inside its folder. better to split it into two files that have clear names inside that folder.
OR (not implemented), remove the folder and just rename the file "Ultra tech meta-traits"