Fix one college limitations for spells
To have a prerequisite of a given level of one-college magery
it needs to specify having the required levels of Magery
that have the right college in the notes,
or doesn't have the one college limit.

A few spells were missing the "doesn't have one college limit" part
so would be satisfiable by one college magery of a different college,
or were missing some or all of their one-college limitations.