Removed Magery's invalid feature (+1/lvl spell lvl)
I thought this +1 per Magery level to your spell level was valid, until a friend told me it wasn't. I looked at the Magery advantage, and double checked through Magic and lightly through Thaumatology. I can't find any mention of this anywhere in the actual books.
So unless someone else knows where this comes from, it should be removed. If I'm right that this is invalid, I'm surprised one of the most commonly used advantages has a mistake like this.
If this actually is valid and comes from a book other than Basic Set, then the advantage should mention this in the page reference section.