Add licenses and certificates from GURPS Traveller Far Trader.
These should be identical to the ones on GURPS Traveller Far Trader
pages 83-85, with some small changes to remove references to skills in
GURPS 3E that are no longer found in 4E.  Shipbuilding has been merged
into Engineer (Starships).  Shipmaster is now Shiphandling (Starship).
This removes the distinction between Limited and Unlimited Master's and
Mate's Licenses, so now there is just one type of Master's License and
one type of Mate's License.

Note that a few of the licenses require a certain number of points
spread across multiple skills, or multiple specializations of a skill
or skills, which are not supported by the current GCS prerequisite
system, so for those licenses that part of the requirement is just
written in the notes rather than actually enforced.  Players and GMs
should manually check that characters actually fulfill any requirements
listed in the notes.