Some changes to allow for latest GCS version (#272)
Equipment file:
Amended all weapons to have +0 parry (except for some of the Glaive ones which retained their +2 parry)

Goblin Emissary:
Removed a "Placeholder" trait I use to make it easier to add new traits to the end in the right group.

Hobgoblin Witch Doctor:
Didn't really make any changes as it all looked OK with no parry or block permitted on the spells.  Saving it seems to have removed those rows from the save file so it shows up as a change.