Ultra Tech Equipment: Correct damage types for Space Biosuit
The damage was previously listed as 15 for percing and cutting, 3 for
everything else. Reflex, and other kinds of 'ballistic' armour tend to
have DR like that.

The listing in Ultra-Tech page 179 explicitly states "use its higher
DR vs. most attacks, but its lower DR against corrosion, crushing,
and toxic damage."

Personally, I think this is incorrect and probably an oversight, as
it's stated to be made of Smart Bioplastic, and Bioplas Armour (page
174) states that it gives full DR to burning and percing, and one third
vs. other damage types, but there's no errata to the contrary so I'll
stick with the book.