Improve prerequisites for Gift of Letters and Gift of Tongues (#342)
* Improve prerequisites for Gift of Letters and Gift of Tongues

The requirement in GURPS Magic is Borrow Language plus three
languages at Accented or better (written for Letters, spoken for
Tongues).  Note that GCS shows languages like "Language: English"
so the right way to check for a language is "starts with: 'Language:'".
(If you miss the colon then the Language Talent advantage is a false
positive.)

Previously the GCS Magic version of these spells just required one
language at Accented or better, which was wrong.  I've changed it to
three languages and comprehension other than None, which is closer but
still not quite correct since Broken comprehension is not supposed to
count.  I don't see a way to express "3 or more languages at Accented or
better" within the current GCS system; it would require nested levels
of boolean logic, or mapping the language levels to numbers.

Dungeon Fantasy RPG doesn't have Broken or Accented languages, so
"3 or more languages with Spoken/Written not set to None" is correct.
(And the alternate clerical prerequisite of Power Investiture 4 is
kept.)

* Tweak the GURPS Magic version of Gift of Letters/Spells

Keep "starts with Language:" (note the colon), which fixes a bug
with Language Talent.

Don't try adding 3 Languages, as GCS doesn't support that idea.

* Remove extra "Language:" from DFRPG Gift of Letters/Tongues

gcs doesn't actually check for multiple distinct Languages when you
list Language 3 times, so this is just misleading.