Loadouts: Move Security Officer's batteries to the Load-Bearing Vest
As-written, it goes in a backpack, but the security officer has no
backpack, only the Spacer Basic Kit's waist pack and the Load-Bearing
Vest.
Given the vest is where batteries for their weapons are meant to go, and
the LBV is part of this template, putting them in the LBV seems more
appropriate than the waists pack.