Martial Arts Skills: add melee attacks to Hammer Fist
The effect of Hammer Fist is that it's a melee attack that does less
damage, but has less risk of injuring the hand.

Melee attack damage for a technique can be represented in gcs, so I
added it here.

Note that this may behave strangely if you have both the brawling-based
Hammer Fist *and* the Karate-based one, but there's very little reason
for a character to have both.