Basic Set Traits: Add a generic Social Stigma
Social Stigma (License Revoked) isn't in any library, and neither
Template Toolkit nor Action define it as something special, so it can be
assumed to be a _generic_ stigma.

Generic Social Stigmas seem to be a -1 to -4 reaction penalty (for -5 to
-20 points) from most people and is easily apparent.

I added equipment modifiers to pick one of the typical ways that your
stigma applies.