Update skils and traits in Spaceship Crew templates to hold source library info (#366)
* Supers: Fix incorrect page reference for Speed Talent

* Space: Add Space Traits

I noticed that there was no library for Code of Honor (Mercenary) and
decided to add all the traits that couldn't be easily represented
without editing traits from existing libraries.

Some of the traits listed in Space have modifiers, too, but they're
stated to be specific versions of Accessibility or Mitigator, so don't
need special treatment.

Strictly-speaking, Social Stigma (Alien) is identical to Minority Group.
Social Stigma (Psionic) is Minority Group, plus an extra licensing
requirement.
Social Stigma (Robot) is Valuable Property, but Space also allows for
Free Machines who have full legal rights.
Social Stigma (Uplifted Animal) is identical to Second-Class Citizen.

* Power-Ups Traits: Add Equipment Bond

This is used in a bunch of places, but not defined in any library.

I have given it a conditional modifier to all uses of a specific piece
of equipment.

* Basic Set Traits: Make Higher Purpose substitutable

i.e. instead of taking the Higher Purpose trait and editing it to say
what the purpose is, you can now substitute a value in when applying it
to your character.

* Basic Set Traits: Add a generic 'Compulsive Behaviour' disadvantage

Basic Set doesn't have any generic ones, TT3: Starship Crew uses
'Compulsive Behaviour (Tinkering)' but this isn't defined anywhere, so
rather than add it to Starship Crew Traits (with no page reference
except to Basic Set's Compulsive Behaviour), I created a generic one
here.

This seems in line with Compulsive Behaviour being a generic trait but
listing a few examples.

* Basic Set Traits: Add a generic Social Stigma

Social Stigma (License Revoked) isn't in any library, and neither
Template Toolkit nor Action define it as something special, so it can be
assumed to be a _generic_ stigma.

Generic Social Stigmas seem to be a -1 to -4 reaction penalty (for -5 to
-20 points) from most people and is easily apparent.

I added equipment modifiers to pick one of the typical ways that your
stigma applies.

* Basic Set Skills: fix Mechanic (@Machine Type@) defaulting from Engineer(Aerospace)

It should default from Engineer of the same machine type.

* Basic Set Skills: add substitutable versions of Fast-Draw, Hazardous Materials, Mathematics and Paleontology

* Template Toolkit: Add Starship Crew Traits

There's one trait described in TT3 that's not added - Jack of All
Trades, but that already exists in Power-Ups Traits from PU3.

* Template Toolkit: Add Starship Crew's Omnicompetent lens

* Template Toolkit: add library sync info to Starship Crew Commander

* Template Toolkit: add library sync info to Helmsman template

* Template Toolkit: add library sync info to Operations Officer template

Also fixed point value of the extra language, made it clear that
Photographic Memory is on the template, and fixed missing/mispriced
skills.

* Template Toolkit: add library sync info to Tactical Officer template

Also fixed the Tactical Officer's Basic Speed

* Template Toolkit: add library sync info to Engineer template

Also:
* makes it clearer that Engineer can have Eidetic Memory (not just
  Photographic Memory)
* Fixed the point value of Computer Programming, Free Fall, Scrounging,
  and Vacc Suit.

* Template Toolkit: add library sync info to Medical Officer template

Also added missing Bioengineering secondary skill and fixed the point
value of Free Fall.

* Template Toolkit: add library sync info to Science Officer template

Also:
* Made it clear both Eidetic and Photographic memory are available.
* Made it clear both Lightning Calculator and Intuitive Mathematician are available.
* Added Better attributes as an option for the Science Officer's 30
  points of advantages.
* Corrected the amount of points in Electronics Operation (sensors),
  Free Fall, NBC Suit and Vacc Suit.

* Template Toolkit: add library sync info to Security Officer template

Also:
* Used specific specialties of Armour and Explosives instead of the
  substitutable one because there's some ambiguity about the names of
  the specialisms.
* added missing Vacc Suit skill.
* Replaced Fast-Draw (@Ammo or Pistol@) with a skill container asking
  for either Fast-Draw/TL (Ammo) or Fast-Draw (Pistol) to avoid confusion
  that Fast-Draw (Pistol) might be TL-specific.
* Fixed the number of points in karate, brawling, boxing, wrestling and
  judo in the multi-role lens.

* Template Toolkit: add library sync info to Loadmaster template

Also:
* Added increased HT to the multi-role lens
* made it clearer that Legendary Loadmaster can have 0.3G or 0.5G improved G-Tolerance

* Template Toolkit: add library sync info to Steward template

Also:
* Split Contact to explicitly be Contact or Contact Group
* Split Cultural Adaptability to make it clearer the Xeno enhancement isn't mandatory.
* Split Electronics Operation (@Comm or Media@) into Electronics Operation
  (Communications) and Electronics Operation (Media) because of ambiguity
  about whether it's called "Comm" or "Communications".