Martial Arts Skills: add melee attacks to Hammer Fist
The effect of Hammer Fist is that it's a melee attack that does less
damage, but has less risk of injuring the hand.

Melee attack damage for a technique can be represented in gcs, so I
added it here.

Because melee attacks need to list which skill (technique in this case)
they default to by name, this requires the technique to be
uniquely-named, so I've included the name of the skill the technique is
based on in its name.