Psionics Traits: Fix tags of talents and make them use tags where possible
I believe that these talents were added at a time when gcs didn't support
adding a skill bonus to any skill with a particular tag.

Adding the skill bonus by tag is also convenient for
pyramid/third-party/homebrew/player-created psionic abilities being
added to existing powers.

This removes the listed individual skills from the talents and gives the
talent to any skill with the tag the talent is named after (i.e.
Psychokinesis Talent adds to all skills tagged Psychokinesis).

It's possible this may cause strange behaviour for some of the talents
with more common-sounding names, like "Control Talent" and "Offense
Talent", but there are no skills tagged "Control" or "Offense" in Basic
Set skills.

As mentioned, it also changed the tags of the Talents, to make sure they
all use tags in the same way.