Use Innate Attack (Projectile) for 4 Missile spells (#423)
Currently we use Projectile for most Missile spells, but Beam
for these 4.

It's not fully clear which to use for Lightning, Explosive Lightning,
and Sunbolt in Magic, but DFRPG Spells specifically says Projectile,
for these, so it makes sense to use that for Magic too.

Spider Silk isn't in DFRPG so we don't have the same precedent there,
but it makes a solid, so I think it clearly should have been
Projectile all along and making it Beam was a mistake.

Making this change unifies all Missile spells under Projectile and
all Jet spells under Beam, which feels right to me, since the skill
should *really* be about how you aim the spell, not its payload.

Co-authored-by: Richard Wilkes <wilkes@me.com>