Add Arden Vul and Traveller calendars.
Arden Vul isn't an official GURPS setting so it should be Home Brew,
but we don't have that concept for Settings/Calendar.  I submitted
it anyway because more examples of calendars are good.

For Traveller I treated the year as one 365-day month called "Day" so
that birthdays would be like "Day 123".  I did not tag Holiday as
outside of any week as I don't think there is support for that, but
it doesn't really matter for current uses of this data.