Basic Set Traits: fix the script in Payload (#441)
* Basic Set Traits: automated changes

These are automatic changes caused by a change to the gcs file format.

* Basic Set Traits: fix the script in Payload

Scripting changes meant that `entity.encumbrance()` changed to
`entity.encumbrance`. In addition, calling
`entity.encumbrance.basicLift` raised an error that overrode the `iff`
block when entity.encumbrance is undefined, so this has been changed to
an ordinary if-else block.